pull 하기전
