+++ lib/data/loading.dart
@@ -1,36 +1,37 @@
-// import 'package:donggu_bat/screens/main_page.dart';
-// import 'package:flutter/material.dart';
-// // import 'my_location.dart';
-// import 'network.dart';
-// // import '../screens/weather_screen.dart';
+import 'package:donggu_bat/screens/main_page.dart';
+import 'package:flutter/material.dart';
+// import 'my_location.dart';
+import 'network.dart';
+// import '../screens/weather_screen.dart';
 
-// const apiKey = '895c7d17476c72440ce44ba845661bbc';
+const apiKey = '895c7d17476c72440ce44ba845661bbc';
 
-// class Loading extends StatelessWidget {
-//   // late double latitude3;
-//   // late double longitude3;
+class Loading extends StatelessWidget {
+  // late double latitude3;
+  // late double longitude3;
+  // dynamic parseWeatherData;
 
-//   dynamic getLocation() async {
-//     // MyLocation myLocation = MyLocation();
-//     // await myLocation.getMyCurrentLocation();
-//     // latitude3 = myLocation.latitude2;
-//     // longitude3 = myLocation.longitude2;
-//     // print(latitude3);
-//     // print(longitude3);
+  dynamic getLocation() async {
+    // MyLocation myLocation = MyLocation();
+    // await myLocation.getMyCurrentLocation();
+    // latitude3 = myLocation.latitude2;
+    // longitude3 = myLocation.longitude2;
+    // print(latitude3);
+    // print(longitude3);
 
-//     Network network = Network('https://api.openweathermap.org/data/2.5/weather?'
-//         'lat=37.550&lon=127.041&exclude=hourly&appid=$apiKey&units=metric');
-//     // https://api.openweathermap.org/data/3.0/onecall?lat=37.550&lon=127.041&exclude=hourly&appid=895c7d17476c72440ce44ba845661bbc&units=metric
+    Network network = Network('https://api.openweathermap.org/data/2.5/weather?'
+        'lat=37.550&lon=127.041&exclude=hourly&appid=$apiKey&units=metric');
+    // https://api.openweathermap.org/data/3.0/onecall?lat=37.550&lon=127.041&exclude=hourly&appid=895c7d17476c72440ce44ba845661bbc&units=metric
 
-//     var weatherData = await network.getJsonData();
-//     print(weatherData);
-//     return MainPage(
-//       parseWeatherData: weatherData,
-//     );
-//   }
+    var weatherData = await network.getJsonData();
+    print(weatherData);
+    print(weatherData['main']['temp']);
+    print(weatherData['name']);
+    return weatherData;
+  }
 
-//   @override
-//   Widget build(BuildContext context) {
-//     return getLocation();
-//   }
-// }
+  @override
+  Widget build(BuildContext context) {
+    return getLocation();
+  }
+}
